+++ .travis.yml
@@ -15,8 +15,9 @@ before_install:
 install:
   - export GIT_FULL_HASH=`git rev-parse HEAD`
   - conda update conda --yes
-  - conda create -n testenv --yes pip nose python=$TRAVIS_PYTHON_VERSION lmfit xraylib numpy scipy scikit-image netcdf4 six coverage
+  - conda create -n testenv --yes pip nose python=$TRAVIS_PYTHON_VERSION xraylib numpy scipy scikit-image netcdf4 six coverage
   - source activate testenv
+  - pip install lmfit==0.8.1
   - python setup.py install
   - pip install coveralls
   - pip install codecov
