+++ MANIFEST.in
@@ -1,2 +1,4 @@
 include versioneer.py
 include skxray/_version.py
+include requirements.txt
+include optional-requirements.txt
